+++ server/src/test/java/io/squashql/js/TestJavascriptLibrary.java
@@ -4,7 +4,7 @@ import io.squashql.jackson.JacksonUtil;
 import io.squashql.query.*;
 import io.squashql.query.builder.Query;
 import io.squashql.query.dto.*;
-import io.squashql.query.join.QueryJoin;
+import io.squashql.query.dto.QueryJoinDto;
 import io.squashql.query.parameter.QueryCacheParameter;
 import io.squashql.util.TestUtil;
 import org.assertj.core.api.Assertions;
@@ -233,13 +233,22 @@ public class TestJavascriptLibrary {
             .select(List.of(b2, c2), List.of(avg("sum", new TableField("f2"))))
             .build();
 
-    QueryJoin join = QueryJoin.from(query1)
+    var c3 = new TableField("myTable3.c3");
+    var query3 = from("myTable3")
+            .select(List.of(c3), List.of(avg("max", new TableField("f3"))))
+            .build();
+
+    QueryJoinDto q = QueryJoinDto.from(query1)
             .join(query2,
                     JoinType.LEFT,
                     all(
                             criterion(b1, b2, ConditionType.EQ),
-                            criterion(c1, c2, ConditionType.EQ)));
-    var q = new QueryJoinDto(join, Map.of(a, new SimpleOrderDto(OrderKeywordDto.ASC)), 10);
+                            criterion(c1, c2, ConditionType.EQ)))
+            .orderBy(Map.of(a, new SimpleOrderDto(OrderKeywordDto.ASC), c3, new SimpleOrderDto(OrderKeywordDto.DESC)))
+            .join(query3, JoinType.INNER)
+            .limit(12);
+
+    q.current = 0;
 
     String name = "build-from-query-join.json"; // The content of this file is generated by the js code.
     QueryJoinDto qjs = JacksonUtil.deserialize(TestUtil.readAllLines(name), QueryJoinDto.class);
