+++ server/src/test/resources/build-from-query-join.json
@@ -1,1 +1,1 @@
-{"first":{"table":{"name":"myTable1","joins":[]},"virtualTableDtos":[],"columns":[{"@class":"io.squashql.query.TableField","fullName":"myTable1.a","tableName":"myTable1","fieldName":"a"},{"@class":"io.squashql.query.TableField","fullName":"myTable1.b","tableName":"myTable1","fieldName":"b"},{"@class":"io.squashql.query.TableField","fullName":"myTable1.c","tableName":"myTable1","fieldName":"c"}],"rollupColumns":[],"columnSets":{},"parameters":{},"measures":[{"@class":"io.squashql.query.AggregatedMeasure","field":{"@class":"io.squashql.query.TableField","fullName":"f1","fieldName":"f1"},"aggregationFunction":"sum","alias":"sum"}],"orders":{},"limit":-1},"second":{"table":{"name":"myTable2","joins":[]},"virtualTableDtos":[],"columns":[{"@class":"io.squashql.query.TableField","fullName":"myTable2.b","tableName":"myTable2","fieldName":"b"},{"@class":"io.squashql.query.TableField","fullName":"myTable2.c","tableName":"myTable2","fieldName":"c"}],"rollupColumns":[],"columnSets":{},"parameters":{},"measures":[{"@class":"io.squashql.query.AggregatedMeasure","field":{"@class":"io.squashql.query.TableField","fullName":"f2","fieldName":"f2"},"aggregationFunction":"avg","alias":"sum"}],"orders":{},"limit":-1},"joinType":"LEFT","joinCondition":{"conditionType":"AND","children":[{"field":{"@class":"io.squashql.query.TableField","fullName":"myTable1.b","tableName":"myTable1","fieldName":"b"},"fieldOther":{"@class":"io.squashql.query.TableField","fullName":"myTable2.b","tableName":"myTable2","fieldName":"b"},"conditionType":"EQ"},{"field":{"@class":"io.squashql.query.TableField","fullName":"myTable1.c","tableName":"myTable1","fieldName":"c"},"fieldOther":{"@class":"io.squashql.query.TableField","fullName":"myTable2.c","tableName":"myTable2","fieldName":"c"},"conditionType":"EQ"}]},"orders":{"{\"@class\":\"io.squashql.query.TableField\",\"fullName\":\"myTable1.a\",\"tableName\":\"myTable1\",\"fieldName\":\"a\"}":{"@class":"io.squashql.query.dto.SimpleOrderDto","order":"ASC"}},"limit":10}
+{"table":{"name":"__cte0__","joins":[{"table":{"name":"__cte1__","joins":[]},"type":"LEFT","joinCriteria":{"conditionType":"AND","children":[{"field":{"@class":"io.squashql.query.TableField","fullName":"myTable1.b","tableName":"myTable1","fieldName":"b"},"fieldOther":{"@class":"io.squashql.query.TableField","fullName":"myTable2.b","tableName":"myTable2","fieldName":"b"},"conditionType":"EQ"},{"field":{"@class":"io.squashql.query.TableField","fullName":"myTable1.c","tableName":"myTable1","fieldName":"c"},"fieldOther":{"@class":"io.squashql.query.TableField","fullName":"myTable2.c","tableName":"myTable2","fieldName":"c"},"conditionType":"EQ"}]}},{"table":{"name":"__cte2__","joins":[]},"type":"INNER"}]},"queries":[{"table":{"name":"myTable1","joins":[]},"virtualTableDtos":[],"columns":[{"@class":"io.squashql.query.TableField","fullName":"myTable1.a","tableName":"myTable1","fieldName":"a"},{"@class":"io.squashql.query.TableField","fullName":"myTable1.b","tableName":"myTable1","fieldName":"b"},{"@class":"io.squashql.query.TableField","fullName":"myTable1.c","tableName":"myTable1","fieldName":"c"}],"rollupColumns":[],"columnSets":{},"parameters":{},"measures":[{"@class":"io.squashql.query.AggregatedMeasure","field":{"@class":"io.squashql.query.TableField","fullName":"f1","fieldName":"f1"},"aggregationFunction":"sum","alias":"sum"}],"orders":{},"limit":-1},{"table":{"name":"myTable2","joins":[]},"virtualTableDtos":[],"columns":[{"@class":"io.squashql.query.TableField","fullName":"myTable2.b","tableName":"myTable2","fieldName":"b"},{"@class":"io.squashql.query.TableField","fullName":"myTable2.c","tableName":"myTable2","fieldName":"c"}],"rollupColumns":[],"columnSets":{},"parameters":{},"measures":[{"@class":"io.squashql.query.AggregatedMeasure","field":{"@class":"io.squashql.query.TableField","fullName":"f2","fieldName":"f2"},"aggregationFunction":"avg","alias":"sum"}],"orders":{},"limit":-1},{"table":{"name":"myTable3","joins":[]},"virtualTableDtos":[],"columns":[{"@class":"io.squashql.query.TableField","fullName":"myTable3.c3","tableName":"myTable3","fieldName":"c3"}],"rollupColumns":[],"columnSets":{},"parameters":{},"measures":[{"@class":"io.squashql.query.AggregatedMeasure","field":{"@class":"io.squashql.query.TableField","fullName":"f3","fieldName":"f3"},"aggregationFunction":"avg","alias":"max"}],"orders":{},"limit":-1}],"orders":{"{\"@class\":\"io.squashql.query.TableField\",\"fullName\":\"myTable1.a\",\"tableName\":\"myTable1\",\"fieldName\":\"a\"}":{"@class":"io.squashql.query.dto.SimpleOrderDto","order":"ASC"},"{\"@class\":\"io.squashql.query.TableField\",\"fullName\":\"myTable3.c3\",\"tableName\":\"myTable3\",\"fieldName\":\"c3\"}":{"@class":"io.squashql.query.dto.SimpleOrderDto","order":"DESC"}},"limit":12}
